+++ packages/@expo/config-types/src/ExpoConfig.ts
@@ -42,7 +42,9 @@ export interface ExpoConfig {
    * The runtime version associated with this manifest.
    * Set this to `{"policy": "nativeVersion"}` to generate it automatically.
    */
-  runtimeVersion?: string | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' };
+  runtimeVersion?:
+    | string
+    | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' | 'fingerprintExperimental' };
   /**
    * Your app version. In addition to this field, you'll also use `ios.buildNumber` and `android.versionCode` — read more about how to version your app [here](https://docs.expo.dev/distribution/app-stores/#versioning-your-app). On iOS this corresponds to `CFBundleShortVersionString`, and on Android, this corresponds to `versionName`. The required format can be found [here](https://developer.apple.com/documentation/bundleresources/information_property_list/cfbundleshortversionstring).
    */
@@ -464,7 +466,9 @@ export interface IOS {
    * The runtime version associated with this manifest for the iOS platform. If provided, this will override the top level runtimeVersion key.
    * Set this to `{"policy": "nativeVersion"}` to generate it automatically.
    */
-  runtimeVersion?: string | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' };
+  runtimeVersion?:
+    | string
+    | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' | 'fingerprintExperimental' };
 }
 /**
  * Configuration that is specific to the Android platform.
@@ -738,7 +742,9 @@ export interface Android {
    * The runtime version associated with this manifest for the Android platform. If provided, this will override the top level runtimeVersion key.
    * Set this to `{"policy": "nativeVersion"}` to generate it automatically.
    */
-  runtimeVersion?: string | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' };
+  runtimeVersion?:
+    | string
+    | { policy: 'nativeVersion' | 'sdkVersion' | 'appVersion' | 'fingerprintExperimental' };
 }
 export interface AndroidIntentFiltersData {
   /**