+++ packages/@expo/cli/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### 💡 Others
 
+- Forward the project root path to `getRuntimeVersionAsync` and `getRuntimeVersionNullableAsync`. ([#24126](https://github.com/expo/expo/pull/24126) by [@mccraveiro](https://github.com/mccraveiro))
+
 ## 0.10.12 — 2023-09-11
 
 ### 🎉 New features